+++ mirrord/cli/src/main.rs
@@ -1,43 +1,40 @@
-use std::{
-    collections::HashMap,
-    fs::File,
-    io::Write,
-    path::{Path, PathBuf},
-    time::Duration,
-};
+#![feature(let_chains)]
 
-use anyhow::{anyhow, Context, Result};
+use std::{collections::HashMap, time::Duration};
+
 use clap::Parser;
 use config::*;
-use const_random::const_random;
 use exec::execvp;
+use execution::MirrordExecution;
+use extension::extension_exec;
+use extract::extract_library;
 use k8s_openapi::api::core::v1::Pod;
 use kube::{api::ListParams, Api};
 use mirrord_auth::AuthConfig;
 use mirrord_config::LayerConfig;
-use mirrord_kube::api::{
-    container::SKIP_NAMES,
-    get_k8s_resource_api,
-    kubernetes::{create_kube_api, KubernetesAPI},
-    AgentManagment,
+use mirrord_kube::{
+    api::{container::SKIP_NAMES, get_k8s_resource_api, kubernetes::create_kube_api},
+    error::KubeApiError,
 };
-use mirrord_operator::client::OperatorApiDiscover;
 use mirrord_progress::{Progress, TaskProgress};
 #[cfg(target_os = "macos")]
 use mirrord_sip::sip_patch;
+use operator::operator_command;
 use semver::Version;
 use serde_json::json;
-use tracing::{debug, error, info, warn};
+use tracing::{error, info, warn};
 use tracing_subscriber::{fmt, prelude::*, registry, EnvFilter};
 
 mod config;
+mod connection;
+mod error;
+mod execution;
+mod extension;
+mod extract;
 mod operator;
 
-#[cfg(target_os = "linux")]
-const INJECTION_ENV_VAR: &str = "LD_PRELOAD";
+pub(crate) use error::{CliError, Result};
 
-#[cfg(target_os = "macos")]
-const INJECTION_ENV_VAR: &str = "DYLD_INSERT_LIBRARIES";
 const PAUSE_WITHOUT_STEAL_WARNING: &str =
     "--pause specified without --steal: Incoming requests to the application will
                                            not be handled. The target container running the deployed application is paused,
@@ -51,109 +48,11 @@ const PAUSE_WITHOUT_STEAL_WARNING: &str =
                                            specifying `--pause`.
     ";
 
-/// For some reason loading dylib from $TMPDIR can get the process killed somehow..?
-#[cfg(target_os = "macos")]
-mod mac {
-    use std::str::FromStr;
-
-    use super::*;
-
-    pub fn temp_dir() -> PathBuf {
-        PathBuf::from_str("/tmp/").unwrap()
-    }
-}
-
-#[cfg(not(target_os = "macos"))]
-use std::env::temp_dir;
-
-#[cfg(target_os = "macos")]
-use mac::temp_dir;
-
-/// Extract to given directory, or tmp by default.
-/// If prefix is true, add a random prefix to the file name that identifies the specific build
-/// of the layer. This is useful for debug purposes usually.
-fn extract_library(
-    dest_dir: Option<String>,
-    progress: &TaskProgress,
-    prefix: bool,
-) -> Result<PathBuf> {
-    let progress = progress.subtask("extracting layer");
-    let extension = Path::new(env!("MIRRORD_LAYER_FILE"))
-        .extension()
-        .unwrap()
-        .to_str()
-        .unwrap();
-
-    let file_name = if prefix {
-        format!("{}-libmirrord_layer.{extension}", const_random!(u64))
-    } else {
-        format!("libmirrord_layer.{extension}")
-    };
-
-    let file_path = match dest_dir {
-        Some(dest_dir) => std::path::Path::new(&dest_dir).join(file_name),
-        None => temp_dir().as_path().join(file_name),
-    };
-    if !file_path.exists() {
-        let mut file = File::create(&file_path)
-            .with_context(|| format!("Path \"{}\" creation failed", file_path.display()))?;
-        let bytes = include_bytes!(env!("MIRRORD_LAYER_FILE"));
-        file.write_all(bytes).unwrap();
-        debug!("Extracted library file to {:?}", &file_path);
-    }
-
-    progress.done_with("layer extracted");
-    Ok(file_path)
-}
-
-fn add_to_preload(path: &str) -> Result<()> {
-    match std::env::var(INJECTION_ENV_VAR) {
-        Ok(value) => {
-            let new_value = format!("{}:{}", value, path);
-            std::env::set_var(INJECTION_ENV_VAR, new_value);
-            Ok(())
-        }
-        Err(std::env::VarError::NotPresent) => {
-            std::env::set_var(INJECTION_ENV_VAR, path);
-            Ok(())
-        }
-        Err(e) => {
-            error!("Failed to set environment variable with error {:?}", e);
-            Err(anyhow!("Failed to set environment variable"))
-        }
-    }
-}
-
+/// Creates an agent and fetches environment variables from it.
+/// Wrapper of async function with tokio for use from sync context.
 #[tokio::main(flavor = "current_thread")]
-async fn create_agent(progress: &TaskProgress) -> Result<()> {
-    let config = LayerConfig::from_env()?;
-
-    if config.operator.enabled
-        && OperatorApiDiscover::discover_operator(&config, progress)
-            .await
-            .is_some()
-    {
-        return Ok(());
-    }
-
-    if config.agent.pause {
-        if config.agent.ephemeral {
-            error!("Pausing is not yet supported together with an ephemeral agent container.");
-            panic!("Mutually exclusive arguments `--pause` and `--ephemeral-container` passed together.");
-        }
-        if !config.feature.network.incoming.is_steal() {
-            warn!("{PAUSE_WITHOUT_STEAL_WARNING}");
-        }
-    }
-    let kube_api = KubernetesAPI::create(&config).await?;
-    let (pod_agent_name, agent_port) = kube_api.create_agent(progress).await?;
-    // Set env var for children to re-use.
-    std::env::set_var("MIRRORD_CONNECT_AGENT", pod_agent_name);
-    std::env::set_var("MIRRORD_CONNECT_PORT", agent_port.to_string());
-
-    // Stop confusion with layer
-    std::env::set_var(mirrord_progress::MIRRORD_PROGRESS_ENV, "off");
-    Ok(())
+async fn start_agent(config: &LayerConfig, progress: &TaskProgress) -> Result<MirrordExecution> {
+    MirrordExecution::start(config, progress).await
 }
 
 fn exec(args: &ExecArgs, progress: &TaskProgress) -> Result<()> {
@@ -267,7 +166,8 @@ fn exec(args: &ExecArgs, progress: &TaskProgress) -> Result<()> {
     if let Some(config_file) = &args.config_file {
         // Set canoncialized path to config file, in case forks/children are in different
         // working directories.
-        let full_path = std::fs::canonicalize(config_file)?;
+        let full_path = std::fs::canonicalize(config_file)
+            .map_err(|e| CliError::ConfigFilePathError(config_file.to_owned(), e))?;
         std::env::set_var("MIRRORD_CONFIG_FILE", full_path);
     }
 
@@ -276,10 +176,6 @@ fn exec(args: &ExecArgs, progress: &TaskProgress) -> Result<()> {
     }
 
     let sub_progress = progress.subtask("preparing to launch process");
-    let library_path = extract_library(args.extract_path.clone(), &sub_progress, true)?;
-    add_to_preload(library_path.to_str().unwrap()).unwrap();
-
-    create_agent(&sub_progress)?;
 
     #[cfg(target_os = "macos")]
     let (_did_sip_patch, binary) = match sip_patch(&args.binary)? {
@@ -290,41 +186,56 @@ fn exec(args: &ExecArgs, progress: &TaskProgress) -> Result<()> {
     #[cfg(not(target_os = "macos"))]
     let binary = args.binary.clone();
 
+    let config = LayerConfig::from_env()?;
+    if config.agent.pause {
+        if config.agent.ephemeral {
+            error!("Pausing is not yet supported together with an ephemeral agent container.");
+            panic!("Mutually exclusive arguments `--pause` and `--ephemeral-container` passed together.");
+        }
+        if !config.feature.network.incoming.is_steal() {
+            warn!("{PAUSE_WITHOUT_STEAL_WARNING}");
+        }
+    }
+
+    let execution_info = start_agent(&config, &sub_progress)?;
+
+    // Stop confusion with layer
+    std::env::set_var(mirrord_progress::MIRRORD_PROGRESS_ENV, "off");
+
+    // Set environment variables from agent + layer settings.
+    for (key, value) in execution_info.environment {
+        std::env::set_var(key, value);
+    }
+
     let mut binary_args = args.binary_args.clone();
     binary_args.insert(0, args.binary.clone());
 
     sub_progress.done_with("ready to launch process");
     // The execve hook is not yet active and does not hijack this call.
-    let err = execvp(binary, binary_args);
+    let err = execvp(binary.clone(), binary_args.clone());
     error!("Couldn't execute {:?}", err);
     #[cfg(all(target_os = "macos", target_arch = "aarch64"))]
     if let exec::Error::Errno(errno::Errno(86)) = err {
         // "Bad CPU type in executable"
         if _did_sip_patch {
-            error!(
-                "The file you are trying to run, {}, is either SIP protected or a script with a
-                shebang that leads to a SIP protected binary. In order to bypass SIP protection,
-                mirrord creates a non-SIP version of the binary and runs that one instead of the
-                protected one. The non-SIP version is however an x86_64 file, so in order to run
-                it on apple hardware, rosetta has to be installed.
-                Rosetta can be installed by runnning:
-
-                softwareupdate --install-rosetta
-
-                ",
-                &args.binary
-            )
+            return Err(CliError::RosettaMissing(binary));
         }
     }
-    Err(anyhow!("Failed to execute binary"))
+    Err(CliError::BinaryExecuteFailed(binary, binary_args))
 }
 
 /// Returns a list of (pod name, [container names]) pairs.
 /// Filtering mesh side cars
 async fn get_kube_pods(namespace: Option<&str>) -> Result<HashMap<String, Vec<String>>> {
-    let client = create_kube_api(None).await?;
+    let client = create_kube_api(None)
+        .await
+        .map_err(CliError::KubernetesApiFailed)?;
     let api: Api<Pod> = get_k8s_resource_api(&client, namespace);
-    let pods = api.list(&ListParams::default()).await?;
+    let pods = api
+        .list(&ListParams::default())
+        .await
+        .map_err(KubeApiError::from)
+        .map_err(CliError::KubernetesApiFailed)?;
 
     // convert pods to (name, container names) pairs
 
@@ -398,7 +309,7 @@ fn cli_progress() -> TaskProgress {
 }
 
 const CURRENT_VERSION: &str = env!("CARGO_PKG_VERSION");
-fn main() -> Result<()> {
+fn main() -> miette::Result<()> {
     registry()
         .with(fmt::layer())
         .with(EnvFilter::from_default_env())
@@ -413,7 +324,8 @@ fn main() -> Result<()> {
         }
         Commands::ListTargets(args) => print_pod_targets(&args)?,
         Commands::Login(args) => login(args)?,
-        Commands::Operator(operator) => operator::operator_command(operator)?,
+        Commands::Operator(args) => operator_command(*args)?,
+        Commands::ExtensionExec(args) => extension_exec(*args)?,
     }
 
     Ok(())